requirements.txt
Flask
Flask_Cors
gradio>=3.7.0
numpy==1.23.0
pyworld==0.2.5
scipy==1.10.0
SoundFile==0.12.1
torch==1.13.1
torchaudio==0.13.1
torchcrepe
tqdm
scikit-maad
praat-parselmouth
onnx
onnxsim
onnxoptimizer
fairseq==0.12.2
librosa==0.9.1
tensorboard
tensorboardX
edge_tts
openai
websocket-client
pyyaml
